+++ packages/react-core/src/core/providers/thirdweb-wallet-provider.tsx
@@ -429,7 +429,7 @@ export function ThirdwebWalletProvider(
               },
             );
           } catch (e) {
-            console.error("Failed to auto connect wallet");
+            console.error("Failed to auto connect personal wallet");
             console.error(e);
             setConnectionStatus("disconnected");
             return;
@@ -460,9 +460,14 @@ export function ThirdwebWalletProvider(
       } catch (e) {
         console.error("Failed to auto connect wallet");
         console.error(e);
-        lastConnectedWalletStorage.removeItem(
-          LAST_CONNECTED_WALLET_STORAGE_KEY,
-        );
+        if (
+          e instanceof Error &&
+          e.message === autoConnectTimeoutErrorMessage
+        ) {
+          lastConnectedWalletStorage.removeItem(
+            LAST_CONNECTED_WALLET_STORAGE_KEY,
+          );
+        }
         setConnectionStatus("disconnected");
       }
     }