+++ packages/react-native/src/evm/components/ConnectWalletFlow/ChooseWallet/ChooseWallet.tsx
@@ -1,7 +1,7 @@
 import { walletIds } from "@thirdweb-dev/wallets";
 import { ModalHeaderTextClose } from "../../base/modal/ModalHeaderTextClose";
 import { WalletConfig } from "@thirdweb-dev/react-core";
-import { ReactNode, useState } from "react";
+import { ReactNode, useMemo, useState } from "react";
 import Box from "../../base/Box";
 import Text from "../../base/Text";
 import ThirdwebLogo from "../../../assets/thirdweb-logo";
@@ -33,7 +33,6 @@ export function ChooseWallet({
   onChooseWallet,
   modalTitleIconUrl,
   onClose,
-  excludeWalletIds = [],
   termsOfServiceUrl,
   privacyPolicyUrl,
 }: ChooseWalletProps) {
@@ -49,14 +48,25 @@ export function ChooseWallet({
       w.id === walletIds.magicLink ||
       (w.id === walletIds.embeddedWallet && w.selectUI),
   );
-  const connectionWallets = wallets
-    .filter(
+
+  const walletsToDisplay = useMemo(() => {
+    const filteredWallets = wallets.filter(
       (wallet) =>
         wallet.id !== walletIds.magicLink &&
         wallet.id !== walletIds.embeddedWallet &&
         wallet.id !== walletIds.localWallet,
-    )
-    .slice(0, 2);
+    );
+
+    const trueItems = filteredWallets.filter(
+      (item) => item.recommended === true,
+    );
+    const falseItems = filteredWallets.filter(
+      (item) => item.recommended !== true,
+    );
+    const sortedWallets = [...trueItems, ...falseItems];
+
+    return sortedWallets;
+  }, [wallets]);
 
   const showToSPrivacyPolicy = termsOfServiceUrl || privacyPolicyUrl;
 
@@ -74,7 +84,7 @@ export function ChooseWallet({
   };
 
   const onSingleWalletPress = () => {
-    onChooseWallet(connectionWallets[0]);
+    onChooseWallet(walletsToDisplay[0]);
   };
 
   const onBackPress = () => {
@@ -135,13 +145,7 @@ export function ChooseWallet({
       {!emailWallet || isConnectAWalletEnabled ? (
         <>
           <ChooseWalletContent
-            wallets={wallets}
-            excludeWalletIds={[
-              ...excludeWalletIds,
-              walletIds.localWallet,
-              walletIds.magicLink,
-              walletIds.embeddedWallet,
-            ]}
+            wallets={walletsToDisplay}
             onChooseWallet={onChooseWallet}
           />
 
@@ -176,7 +180,7 @@ export function ChooseWallet({
       ) : null}
       {emailWallet &&
       !isConnectAWalletEnabled &&
-      (guestWallet || connectionWallets.length > 0) ? (
+      (guestWallet || walletsToDisplay.length > 0) ? (
         <Box
           mt="md"
           marginHorizontal="xl"
@@ -193,15 +197,15 @@ export function ChooseWallet({
       ) : null}
       {emailWallet &&
       !isConnectAWalletEnabled &&
-      connectionWallets.length > 0 ? (
-        connectionWallets.length === 1 ? (
+      walletsToDisplay.length > 0 ? (
+        walletsToDisplay.length === 1 ? (
           <WalletButton
             marginHorizontal="xl"
             paddingHorizontal="none"
             paddingVertical="none"
             mt="md"
-            walletIconUrl={connectionWallets[0].meta.iconURL}
-            name={connectionWallets[0].meta.name}
+            walletIconUrl={walletsToDisplay[0].meta.iconURL}
+            name={walletsToDisplay[0].meta.name}
             onPress={onSingleWalletPress}
           />
         ) : (
@@ -217,7 +221,10 @@ export function ChooseWallet({
             borderWidth={1}
             onPress={onConnectAWalletPress}
           >
-            {connectionWallets.map((wallet) => {
+            {walletsToDisplay.map((wallet, index) => {
+              if (index > 1) {
+                return null;
+              }
               return (
                 <Box key={wallet.meta.name} mr="xxs">
                   <ImageSvgUri