+++ packages/react-native/src/evm/components/ConnectWalletFlow/ChooseWallet/ChooseWalletContent.tsx
@@ -1,4 +1,3 @@
-import { useMemo } from "react";
 import type { WalletConfig } from "@thirdweb-dev/react-core";
 import { Dimensions, ScrollView, View } from "react-native";
 import { WalletButton } from "../../base/WalletButton";
@@ -7,7 +6,6 @@ import { useTheme } from "@shopify/restyle";
 
 interface ChooseWalletContentProps {
   wallets: WalletConfig[];
-  excludeWalletIds?: string[];
   onChooseWallet: (wallet: WalletConfig, data?: any) => void;
 }
 
@@ -15,26 +13,9 @@ const MAX_HEIGHT = Dimensions.get("window").height * 0.3;
 
 export const ChooseWalletContent = ({
   wallets,
-  excludeWalletIds,
   onChooseWallet,
 }: ChooseWalletContentProps) => {
   const theme = useTheme();
-
-  const walletsToDisplay = useMemo(() => {
-    const filteredWallets = wallets.filter(
-      (w) => !!!excludeWalletIds?.find((ewId) => ewId === w.id),
-    );
-
-    const trueItems = filteredWallets.filter(
-      (item) => item.recommended === true,
-    );
-    const falseItems = filteredWallets.filter(
-      (item) => item.recommended !== true,
-    );
-    const sortedWallets = [...trueItems, ...falseItems];
-
-    return sortedWallets;
-  }, [wallets, excludeWalletIds]);
 
   return (
     <View style={{ flexDirection: "column", maxHeight: MAX_HEIGHT }}>
@@ -45,9 +26,8 @@ export const ChooseWalletContent = ({
           paddingHorizontal: 16,
         }}
       >
-        {walletsToDisplay.map((item, index) => {
-          const marginBottom =
-            index === walletsToDisplay.length - 1 ? "none" : "xxs";
+        {wallets.map((item, index) => {
+          const marginBottom = index === wallets.length - 1 ? "none" : "xxs";
           return (
             <Box key={item.id}>
               {item.selectUI ? (